requirements.txt
fastapi
uvicorn
SQLAlchemy
passlib
python-jose
python-multipart
pymysql
pydantic[email]
cryptography
python-dotenv
psycopg2-binary